+++ src/event.js
@@ -44,6 +44,7 @@ chrome.storage.sync.get(null,function(items){
 });
 
 chrome.runtime.onMessage.addListener(function(message,sender,sendResponse){
+	console.log(message);
 	if ((message.type=='dataslayer_gtm')||(message.type=='dataslayer_tlm')){
 		message.tabID=sender.tab.id;
 		devtoolsPort.forEach(function(v,i,x){